% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_class.R
\name{AIC.gips}
\alias{AIC.gips}
\alias{BIC.gips}
\title{Akaike's An Information Criterion for gips class}
\usage{
\method{AIC}{gips}(object, ..., k = 2)

\method{BIC}{gips}(object, ...)
}
\arguments{
\item{object}{An object of class "gips"; usually a result of a \code{\link[=find_MAP]{find_MAP()}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Akaike's An Information Criterion for gips class
}
\section{Functions}{
\itemize{
\item \code{BIC(gips)}: 

}}
\examples{
S <- matrix(c(5.15,2.05,3.10,1.99,
              2.05,5.09,2.03,3.07,
              3.10,2.03,5.21,1.97,
              1.99,3.07,1.97,5.13), nrow = 4)
g <- gips(S, 14)
AIC(g) # 254

g_map <- find_MAP(g, optimizer = "brute_force")
AIC(g_map) # 240 < 254, so this is better
================================================================================
BIC(g) # 261
BIC(g_map) # 242 < 261, so this is better
}
