% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_class.R
\name{as.character.gips}
\alias{as.character.gips}
\title{Transform \code{gips} object to character vector}
\usage{
\method{as.character}{gips}(x, ...)
}
\arguments{
\item{x}{An object of a \code{gips} class.}

\item{...}{Further arguments passed to \code{\link[=as.character.gips_perm]{as.character.gips_perm()}}.}
}
\value{
Returns an object of a \code{character} type.
}
\description{
Implementation of S3 method.
}
\examples{
A <- matrix(rnorm(4 * 4), nrow = 4)
S <- t(A) \%*\% A
g <- gips(S, 14, perm = "(123)")
as.character(g)
}
\seealso{
\code{\link[=as.character.gips_perm]{as.character.gips_perm()}}

\code{\link[permutations:print]{permutations::as.character.cycle()}}
}
