% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_perm_class.R
\name{as.character.gips_perm}
\alias{as.character.gips_perm}
\title{Transform \code{gips_perm} object to character vector}
\usage{
\method{as.character}{gips_perm}(x, ...)
}
\arguments{
\item{x}{An object of a \code{gips_perm} class.}

\item{...}{Further arguments passed to \code{\link[permutations:print]{permutations::as.character.cycle()}}.}
}
\value{
Returns an object of a \code{character} type.
}
\description{
Implementation of S3 method.
}
\section{Methods (by class)}{
\itemize{
\item \code{as.character(gips_perm)}: 

}}
\examples{
g_perm <- gips_perm(permutations::as.cycle("(5,4)"), 5)
as.character(g_perm)
}
\seealso{
\code{\link[permutations:print]{permutations::as.character.cycle()}}
}
