% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algorithms.R
\name{best_growth}
\alias{best_growth}
\title{Best growth algorithm}
\usage{
best_growth(
  U,
  n_number,
  max_iter = 5,
  start_perm = NULL,
  delta = 3,
  D_matrix = NULL,
  show_progress_bar = TRUE
)
}
\arguments{
\item{U}{matrix, sum of outer products of data. \code{U} = sum(t(Z) \%*\% Z), where Z is observed data.}

\item{n_number}{number of data points that \code{U} is based on.}

\item{max_iter}{number of iterations for an algorithm to perform. Default 5. At least 2. Can be Inf.}

\item{start_perm}{starting permutation for the algorithm; an element of class "cycle". When NULL, identity permutation is taken.}

\item{delta}{hyper-parameter of a Bayesian model. Has to be bigger than 2.}

\item{D_matrix}{hyper-parameter of a Bayesian model. Square matrix of the same size as \code{U}. When NULL, the identity matrix is taken.}

\item{show_progress_bar}{boolean, indicate weather or not show the progress bar. \code{show_progress_bar == TRUE} is not supported for \code{max_iter == Inf}.}
}
\value{
object of class gips; list of 10 items: \code{acceptance_rate},
\code{goal_function_logvalues} - of size \code{iterations_performed} + 1, \code{points}, \code{found_point},
\code{found_point_function_logvalue}, \code{last_point},
\code{last_point_function_logvalue}, \code{iterations_performed}, \code{U_used},
\code{did_converged} - indicates if the algorithm converged.
}
\description{
Uses best growth algorithm to find the permutation that maximizes the likelihood of observed data.
}
\examples{
perm_size <- 6
mu <- numeric(perm_size)
# sigma is a matrix invariant under permutation (1,2,3,4,5,6)
sigma_matrix <- matrix(data = c(1.0, 0.8, 0.6, 0.4, 0.6, 0.8,
                                0.8, 1.0, 0.8, 0.6, 0.4, 0.6,
                                0.6, 0.8, 1.0, 0.8, 0.6, 0.4,
                                0.4, 0.6, 0.8, 1.0, 0.8, 0.6,
                                0.6, 0.4, 0.6, 0.8, 1.0, 0.8,
                                0.8, 0.6, 0.4, 0.6, 0.8, 1.0),
                       nrow=perm_size, byrow = TRUE)
n_number <- 13
Z <- MASS::mvrnorm(n_number, mu = mu, Sigma = sigma_matrix)
U <- (t(Z) \%*\% Z)
start_perm <- permutations::id
bg <- best_growth(U=U, n_number=n_number, max_iter=10, start_perm=start_perm,
                  show_progress_bar=FALSE) # Algorithm did converge in 4 iterations
plot(bg)
}
