% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gamma_function.R
\name{calculate_gamma_function}
\alias{calculate_gamma_function}
\title{Calculate Gamma function}
\usage{
calculate_gamma_function(perm, lambda)
}
\arguments{
\item{perm}{an element of class "cycle"}

\item{lambda}{positive real number}
}
\value{
Value of Gamma function
}
\description{
Theorem 8 from the paper, using the formula (19) from the paper
}
