% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gamma_function.R
\name{calculate_gamma_omega}
\alias{calculate_gamma_omega}
\title{Calculate single Gamma omega function}
\usage{
calculate_gamma_omega(lambda, dim_omega_i, r_i, d_i)
}
\arguments{
\item{lambda}{positive real number}

\item{dim_omega_i}{single element from \code{get_structure_constants}}

\item{r_i}{single element  from \code{get_structure_constants}}

\item{d_i}{single element  from \code{get_structure_constants}}
}
\value{
Logarithm of value of Gamma function
}
\description{
Using the formula (12) from the paper
}
