% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_posteriori_of_gips.R
\name{compare_posteriories_of_perms}
\alias{compare_posteriories_of_perms}
\alias{compare_log_posteriories_of_perms}
\title{Compare the posteriori probabilities of 2 permutations}
\usage{
compare_posteriories_of_perms(
  perm1,
  perm2 = "()",
  S = NULL,
  number_of_observations = NULL,
  delta = 3,
  D_matrix = NULL,
  was_mean_estimated = TRUE,
  print_output = TRUE
)

compare_log_posteriories_of_perms(
  perm1,
  perm2 = "()",
  S = NULL,
  number_of_observations = NULL,
  delta = 3,
  D_matrix = NULL,
  was_mean_estimated = TRUE,
  print_output = TRUE
)
}
\arguments{
\item{perm1, perm2}{Permutations to compare. They will be compared
how many times \code{perm1} is more likely than \code{perm2}.
Those can be provided as the \code{gips} object,
the \code{gips_perm} object or anything that can be used as
the \code{x} parameter in the \code{\link[=gips_perm]{gips_perm()}} function.
They do not have to be of the same class.}

\item{S, number_of_observations, delta, D_matrix, was_mean_estimated}{The same parameters as in the \code{\link[=gips]{gips()}} function.
If at least one of \code{perm1} or \code{perm2} is of a \code{gips} class,
they are overwritten.}

\item{print_output}{A logical (TRUE or FALSE).
When TRUE, the computed value will be printed with
additional text and returned invisibly. When FALSE,
the computed value will be returned visibly.}
}
\value{
\code{compare_posteriories_of_perms} returns the value of
how many times the \code{perm1} is more likely than \code{perm2}.

\code{compare_log_posteriories_of_perms} returns the logarithm of
how many times the \code{perm1} is more likely than \code{perm2}.
}
\description{
Check which permutation is more likely and how much more likely.
}
\section{Functions}{
\itemize{
\item \code{compare_log_posteriories_of_perms()}: More stable,
logarithmic version of \code{compare_posteriories_of_perms}.
The natural logarithm is used.

}}
\examples{
require("MASS") # for mvrnorm()

perm_size <- 6
mu <- runif(6, -10, 10) # Assume we don't know the mean
sigma_matrix <- matrix(
  data = c(
    1.0, 0.8, 0.6, 0.4, 0.6, 0.8,
    0.8, 1.0, 0.8, 0.6, 0.4, 0.6,
    0.6, 0.8, 1.0, 0.8, 0.6, 0.4,
    0.4, 0.6, 0.8, 1.0, 0.8, 0.6,
    0.6, 0.4, 0.6, 0.8, 1.0, 0.8,
    0.8, 0.6, 0.4, 0.6, 0.8, 1.0
  ),
  nrow = perm_size, byrow = TRUE
) # sigma_matrix is a matrix invariant under permutation (1,2,3,4,5,6)
number_of_observations <- 13
Z <- MASS::mvrnorm(number_of_observations, mu = mu, Sigma = sigma_matrix)
S <- cov(Z) # Assume we have to estimate the mean

g <- gips(S, number_of_observations)
g_map <- find_MAP(g, max_iter = 10, show_progress_bar = FALSE, optimizer = "MH")

compare_posteriories_of_perms(g_map, g, print_output = FALSE)
compare_log_posteriories_of_perms(g_map, g, print_output = FALSE)
}
\seealso{
\itemize{
\item \code{\link[=print.gips]{print.gips()}} - The function that prints the posterior of
the optimized \code{gips} object compared to the starting permutation.
\item \code{\link[=summary.gips]{summary.gips()}} - The function that calculates the posterior of
the optimized \code{gips} object compared to the starting permutation.
}
}
