% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_representative.R
\name{get_coprimes}
\alias{get_coprimes}
\title{Get Coprime numbers}
\usage{
get_coprimes(n)
}
\arguments{
\item{n}{single integer}
}
\value{
all integers smaller than n, that are coprime to n
}
\description{
Get Coprime numbers
}
