% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure_constants.R
\name{get_cycle_represesentatives_and_lenghts}
\alias{get_cycle_represesentatives_and_lenghts}
\title{Get cycle representatives and lengths}
\usage{
get_cycle_represesentatives_and_lenghts(perm)
}
\arguments{
\item{perm}{an element of class "cycle"}
}
\value{
list with 2 items: 'representatives' and 'cycle_lengths'
}
\description{
Essentially get iC, pC from paper
}
