% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_representative.R
\name{get_representative}
\alias{get_representative}
\title{Get Representative of a cyclic permutation group}
\usage{
get_representative(perm)
}
\arguments{
\item{perm}{an element of class "cycle"}
}
\description{
Essentially a "nu" function from paper (beginning of section 4.1.1)
\verb{get_representative(perm) == nu(< perm >)}, where "\code{nu}" is from the paper, and "\verb{< perm >}" is a cyclic group generated by the permutation
}
