% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure_constants.R
\name{get_structure_constants}
\alias{get_structure_constants}
\title{Get Structure Constants}
\usage{
get_structure_constants(perm)
}
\arguments{
\item{perm}{an element of class "cycle"}
}
\value{
list of 4 items: \code{r}, \code{d}, \code{k}, \code{L} - vectors of constants from theorem 1
}
\description{
Get Structure Constants from theorem 5 of the paper
}
\examples{
perm <- permutations::as.cycle(permutations::as.word(c(1,2,3,5,4)))
get_structure_constants(perm)
}
