% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_of_gips.R
\name{log_likelihood_of_gips}
\alias{log_likelihood_of_gips}
\title{The log likelihood that the CoV matrix is invariant under permutation.}
\usage{
log_likelihood_of_gips(g)
}
\arguments{
\item{g}{object of class \code{gips}}
}
\description{
To be more precise, it is the logarithm of an unnormalized posterior probability.
Calculated according to equation (33) and (27) from the paper. If \code{Inf} or \code{NaN} is reached, produces a warning.
It is the goal function for optimization algorithms. The \code{perm_proposal} that maximizes this function is the maximum likelihood estimator.
}
\examples{
c <- permutations::as.cycle(permutations::as.word(c(2,1)))
S1 <- matrix(c(1,0.5,0.5,2), nrow=2, byrow = TRUE)
g <- gips(S1, 100, perm=c)
log_likelihood_of_gips(g)
}
