% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_S3_functions.R
\name{plot.MH}
\alias{plot.MH}
\title{Plot Metropolis-Hasting convergence}
\usage{
\method{plot}{MH}(
  x,
  logarithmic = TRUE,
  type = "max",
  title_text = "Convergence plot",
  xlabel = "number of function calls",
  ylabel = NULL,
  ...
)
}
\arguments{
\item{x}{object of class MH. Output of MH function.}

\item{logarithmic}{boolean.}

\item{type}{To plot the maximum found value or all found values.}

\item{title_text}{Text to be in a title of the plot.}

\item{xlabel}{Text to be on the bottom of the plot.}

\item{ylabel}{Text to be on the left of the plot.}

\item{...}{additional arguments passed to \code{\link{print}}.}
}
\value{
Invisible NULL.
}
\description{
Plot method for MH objects returned by \code{\link{MH}}.
}
