% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_functions.R
\name{plot.optimized_best_growth}
\alias{plot.optimized_best_growth}
\title{Plot optimization convergence for \code{best_growth} algorithm}
\usage{
\method{plot}{optimized_best_growth}(
  x,
  logarithmic = TRUE,
  title_text = "Convergence plot",
  xlabel = "number of function calls",
  ylabel = NULL,
  show_legend = TRUE,
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{Object of class gips.}

\item{logarithmic}{boolean.}

\item{title_text}{Text to be in a title of the plot.}

\item{xlabel}{Text to be on the bottom of the plot.}

\item{ylabel}{Text to be on the left of the plot.}

\item{show_legend}{boolean.}

\item{ylim}{Limits of y axis. When \code{NULL}, the minimum and maximum of the \code{\link{goal_function}} is taken.}

\item{...}{additional arguments passed to \code{\link{print}}.}
}
\value{
Invisible NULL.
}
\description{
Plot method for gips objects.
}
