% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_functions.R
\name{print.gips}
\alias{print.gips}
\title{Printing gips object}
\usage{
\method{print}{gips}(x, log_value = FALSE, ...)
}
\arguments{
\item{x}{object of class gips.}

\item{log_value}{logical. Weather to print the exp of a value of a \code{\link{goal_function}} or leave it in logarithmic form.}

\item{...}{additional arguments passed to \code{\link{cat}}.}
}
\value{
Invisible NULL.
}
\description{
Printing function for gips class.
}
