% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_perm_class.R
\name{print.gips_perm}
\alias{print.gips_perm}
\title{Print gips_perm}
\usage{
\method{print}{gips_perm}(x, ...)
}
\arguments{
\item{x}{An object of \code{gips_perm} class.}

\item{...}{Further arguments passed to \code{\link[permutations]{print.cycle}}.}
}
\description{
Implementation of S3 method.
}
