% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_S3_functions.R
\name{print.MH}
\alias{print.MH}
\title{Printing MH}
\usage{
\method{print}{MH}(x, ...)
}
\arguments{
\item{x}{object of class MH. Output of MH function.}

\item{...}{additional arguments passed to \code{\link{cat}}.}
}
\description{
Printing function for MH.
}
