% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_posteriori_of_gips.R
\name{runif_transposition}
\alias{runif_transposition}
\title{Uniformly random transposition of perm_size elements}
\usage{
runif_transposition(perm_size)
}
\arguments{
\item{perm_size}{size from which take transpositions}
}
\description{
Uniformly random transposition of perm_size elements
}
