% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_class.R
\name{summary.gips}
\alias{summary.gips}
\title{Summarizing the gips object}
\usage{
\method{summary}{gips}(object, ...)
}
\arguments{
\item{object}{An object of class "gips", usually, a result of a \code{\link[=find_MAP]{find_MAP()}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The function \code{summary.gips} computes and returns a list of summary
statistics of the \code{gips} given in object. Those are: TODO()
}
\description{
\code{summary} method for class "gips".
}
\seealso{
\code{\link[=project_matrix]{project_matrix()}}, \code{\link[=find_MAP]{find_MAP()}}
}
