% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metropolis_Hastings.R
\name{test_goal_function}
\alias{test_goal_function}
\title{example goal function
Used just for testing}
\usage{
test_goal_function(perm_proposal)
}
\arguments{
\item{perm_proposal}{permutation of interest}
}
\description{
example goal function
Used just for testing
}
