% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_perm_class.R
\name{validate_gips_perm}
\alias{validate_gips_perm}
\title{Validate gips_perm object}
\usage{
validate_gips_perm(g)
}
\arguments{
\item{g}{Element to be checked if it is proper element of class \code{gips_perm}.}
}
\description{
Validate gips_perm object
}
