% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_class.R
\name{AIC.gips}
\alias{AIC.gips}
\alias{BIC.gips}
\title{Akaike's An Information Criterion for gips class}
\usage{
\method{AIC}{gips}(object, ..., k = 2)

\method{BIC}{gips}(object, ...)
}
\arguments{
\item{object}{An object of class "gips"; usually a result of a \code{\link[=find_MAP]{find_MAP()}}.}

\item{...}{Further arguments will be ignored}

\item{k}{numeric, the \emph{penalty} per parameter to be used; the
    default \code{k = 2} is the classical AIC.}
}
\description{
Akaike's An Information Criterion for gips class
}
\section{Functions}{
\itemize{
\item \code{BIC(gips)}: Schwarz's Bayesian Information Criterion

}}
\examples{
S <- matrix(c(5.15,2.05,3.10,1.99,
              2.05,5.09,2.03,3.07,
              3.10,2.03,5.21,1.97,
              1.99,3.07,1.97,5.13), nrow = 4)
g <- gips(S, 14)
AIC(g) # 238

g_map <- find_MAP(g, optimizer = "brute_force")
AIC(g_map) # 224 < 238, so g_map is better than g in AIC
# ================================================================================
BIC(g) # 244
BIC(g_map) # 226 < 244, so g_map is better than g in BIC
}
\seealso{
\itemize{
\item \code{\link[=AIC]{AIC()}}, \code{\link[=BIC]{BIC()}} - Generic functions
this \code{\link[=AIC.gips]{AIC.gips()}} and \code{\link[=BIC.gips]{BIC.gips()}} extend.
\item \code{\link[=find_MAP]{find_MAP()}} - Usually, the \code{AIC.gips()} and \code{BIC.gips()}
are called on the output of \code{find_MAP()}.
\item \code{\link[=logLik.gips]{logLik.gips()}} - Calculates the log likelihood for
the \code{gips} object. Important part of the Information Criteria.
}
}
