% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_perm_class.R
\name{as.character.gips_perm}
\alias{as.character.gips_perm}
\title{Coerce gips_perm to character vector}
\usage{
\method{as.character}{gips_perm}(x, ...)
}
\arguments{
\item{x}{Object of \code{gips_perm} class}

\item{...}{further arguments passed to \code{\link[base]{as.character}}}
}
\description{
Implementation of S3 method.
}
