% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gamma_function.R
\name{calculate_gamma_function}
\alias{calculate_gamma_function}
\title{Calculate Gamma function}
\usage{
calculate_gamma_function(perm, lambda)
}
\arguments{
\item{perm}{An object of class \code{gips_perm}.}

\item{lambda}{A positive real number.}
}
\value{
Value of Gamma function.
}
\description{
It calculates the value of the integral defined in
\href{https://doi.org/10.1214/22-AOS2174}{Definition 11 from references}.
It is implementation of the
\href{https://doi.org/10.1214/22-AOS2174}{Theorem 8 from references}
and is using the
\href{https://doi.org/10.1214/22-AOS2174}{formula (19) from references}.
}
\examples{
id_perm <- gips_perm(permutations::id, 2)
calculate_gamma_function(id_perm, 0.5001)  # 10.7...
calculate_gamma_function(id_perm, 0.50000001)  # 19.9...
calculate_gamma_function(id_perm, 0.500000000001)  # 29.1...
# calculate_gamma_function(id_perm, 0.5) # integral diverges; returns Inf and warning
}
\references{
Piotr Graczyk, Hideyuki Ishi, Bartosz Kolodziejek, Hélène Massam.
"Model selection in the space of Gaussian models invariant by symmetry."
The Annals of Statistics, 50(3) 1747-1774 June 2022.
\href{https://doi.org/10.1214/22-AOS2174}{DOI: 10.1214/22-AOS2174}
}
\seealso{
\itemize{
\item \code{\link[=get_structure_constants]{get_structure_constants()}} - The function useful inside
the \code{calculate_gamma_function()} function.
\item \code{\link[=log_posteriori_of_gips]{log_posteriori_of_gips()}} - The function that uses
the values of the gamma function calculable
with \code{calculate_gamma_function()}.
}
}
