% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gamma_function.R
\name{calculate_gamma_function}
\alias{calculate_gamma_function}
\title{Calculate Gamma function}
\usage{
calculate_gamma_function(perm, lambda)
}
\arguments{
\item{perm}{an element of class "gips_perm"}

\item{lambda}{positive real number}
}
\value{
Value of Gamma function
}
\description{
Theorem 8 from the paper, using the formula (19) from the paper
}
\examples{
id_perm <- gips_perm(permutations::id, 2)
calculate_gamma_function(id_perm, 0.5001)
calculate_gamma_function(id_perm, 0.50000001)
calculate_gamma_function(id_perm, 0.500000000001)
#calculate_gamma_function(id_perm, 0.5) # integral diverges; returns Inf and warning
}
