% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gamma_function.R
\name{G_function}
\alias{G_function}
\title{G_function for goal_function}
\usage{
G_function(perm, structure_constants, delta = 3)
}
\arguments{
\item{perm}{permutation of interest, object of \code{cycle} class}

\item{structure_constants}{constants from \code{get_structure_constants} function}

\item{delta}{parameter of a method}
}
\value{
Sum of logarithms of elements of \code{calculate_gamma_omega} from i to L.
It is log of a product part of equation (27). For more information, see Issue #3 on \code{gips}' GitHub
}
\description{
G_function for goal_function
}
\examples{
perm_size <- 6
perm <- permutations::as.cycle(permutations::as.word(c(2,3,1,5,4,6)))
structure_constants <- get_structure_constants(perm, perm_size)
gips:::G_function(perm, structure_constants, 3)
}
