% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure_constants.R
\name{get_structure_constants}
\alias{get_structure_constants}
\title{Get Structure Constants}
\usage{
get_structure_constants(perm)
}
\arguments{
\item{perm}{An object of class \code{gips_perm}.}
}
\value{
A list of 5 items: \code{r}, \code{d}, \code{k}, \code{L}, \code{dim_omega} - vectors of constants from theorem 1 and beginning of section 3.1.
}
\description{
Get Structure Constants from theorem 5 of the paper
}
\examples{
perm <- gips_perm(permutations::as.word(c(1, 2, 3, 5, 4)), 5)
get_structure_constants(perm)
}
\seealso{
\code{\link[=calculate_gamma_function]{calculate_gamma_function()}}
}
