% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_perm_class.R
\name{gips_perm}
\alias{gips_perm}
\alias{new_gips_perm}
\title{Permutation object}
\usage{
gips_perm(x, size)

new_gips_perm(rearranged_cycles, size)
}
\arguments{
\item{x}{An object created with \code{permutations} package, or any object that
can be coerced using \code{\link[permutations:permutation]{permutations::permutation()}} function.}

\item{size}{An integer. Size of permutation (AKA cardinality of set, on which permutation
is defined).}

\item{rearranged_cycles}{A list of rearranged integer vectors. Each vector corresponds to a single cycle
of a permutation.}
}
\description{
Create permutation objects to be passed to functions of \code{gips} package.
}
\section{Functions}{
\itemize{
\item \code{new_gips_perm()}: Constructor

Only intended for low-level use.

}}
\examples{
gperm <- gips_perm(permutations::as.word(c(1, 2, 3, 5, 4)), 5)
gperm <- gips_perm(permutations::as.cycle("(5,4)"), 5)
# note the necessity of `size` parameter
gperm <- gips_perm(permutations::as.cycle("(5,4)"), 7)
gperm <- gips_perm("(1,2)(5,4)", 7)
gperm

}
\seealso{
\code{\link[permutations:permutation]{permutations::permutation()}}, \code{\link[=gips]{gips()}}
}
