% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goal_function.R
\name{goal_function}
\alias{goal_function}
\title{The goal function for optimization functions.}
\usage{
goal_function(perm_proposal, n_number, U, delta = 3, D_matrix = NULL)
}
\arguments{
\item{perm_proposal}{permutation of interest.}

\item{n_number}{number of random variables that \code{U} is based on.}

\item{U}{matrix that the projection of is wanted.}

\item{delta}{hyper-parameter of a Bayesian model. Has to be bigger than 2.}

\item{D_matrix}{hyper-parameter of a Bayesian model. Square matrix of size \code{perm_size}. When NULL, the identity matrix is taken.}
}
\description{
Calculate the logarithm of function proportional to a posteriori distribution, according to equation (33) and (27). If \code{Inf} or \code{NaN} is reached, produces a warning.
}
\examples{
c <- permutations::as.cycle(permutations::as.word(c(2,1)))
U1 <- matrix(c(1,0.5,0.5,2), nrow=2,byrow = TRUE)
goal_function(c, 100, U1)
}
