% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metropolis_Hastings.R
\name{goal_function}
\alias{goal_function}
\title{goal function for MH}
\usage{
goal_function(perm_proposal, perm_size, n, U, delta = 3, D_matrix = NULL)
}
\arguments{
\item{perm_proposal}{Permutation of interest}

\item{perm_size}{size of a permutation}

\item{n}{Size of a sample}

\item{U}{Matrix that the projection of is wanted}

\item{delta}{hyper-parameter of a Bayesian model. Has to be bigger than 2.}

\item{D_matrix}{hyper-parameter of a Bayesian model. Square matrix of size \code{perm_size}. When NULL, the identity matrix is taken}
}
\description{
If infinite value is reached, produces a warning
}
