% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_of_perm.R
\name{log_likelihood_of_perm}
\alias{log_likelihood_of_perm}
\title{The log likelihood that the CoV matrix is invariant under permutation.}
\usage{
log_likelihood_of_perm(
  perm_proposal,
  number_of_observations,
  S,
  delta = 3,
  D_matrix = NULL
)
}
\arguments{
\item{perm_proposal}{permutation of interest.}

\item{number_of_observations}{number of random variables that \code{S} is based on.}

\item{S}{matrix, estimated covariance matrix. When Z is observed data: \code{S = sum(t(Z) \%*\% Z)/number_of_observations}, if one know the theoretical mean is 0; # TODO(What if one have to estimate the theoretical mean with the empirical mean)}

\item{delta}{hyper-parameter of a Bayesian model. Has to be bigger than 2.}

\item{D_matrix}{hyper-parameter of a Bayesian model. Square matrix of size \code{perm_size}. When NULL, the identity matrix is taken.}
}
\description{
Calculate the logarithm of function proportional to a posteriori distribution, according to equation (33) and (27). If \code{Inf} or \code{NaN} is reached, produces a warning.
Keep in mind this is not the log likelihood per se, but rather the funcition that is proportional to the log likelihood.
It is the goal function for optimization algorithms.
}
\examples{
c <- permutations::as.cycle(permutations::as.word(c(2,1)))
S1 <- matrix(c(1,0.5,0.5,2), nrow=2, byrow = TRUE)
log_likelihood_of_perm(c, 100, S1)
}
