% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metropolis_Hastings.R
\name{MH}
\alias{MH}
\title{Metropolis-Hastings algorithm}
\usage{
MH(start, p, max_iter)
}
\arguments{
\item{start}{Start of the algorithm; an element of class "cycle"}

\item{p}{The dimension of interest}

\item{max_iter}{Number of iterations}
}
\value{
list of 3 items: \code{acceptance_rate}, \code{goal_function_values}, \code{points}
}
\description{
Perform the algorithm on the permutations
}
\examples{
start <- as.cycle(as.word(c(2,3,1,4,6,5,7,8)))
mh <- MH(start = start, 8, 100)
}
