% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metropolis_Hastings.R
\name{MH}
\alias{MH}
\title{Metropolis-Hastings algorithm}
\usage{
MH(
  U,
  n_number,
  max_iter,
  start = NULL,
  delta = 3,
  D_matrix = NULL,
  show_progress_bar = TRUE
)
}
\arguments{
\item{U}{matrix, sum of outer products of data. \code{U} = sum(t(Z) \%*\% Z), where Z is observed data.}

\item{n_number}{number of data points that \code{U} is based on.}

\item{max_iter}{number of iterations for an algorithm to perform.}

\item{start}{starting permutation for the algorithm; an element of class "cycle". When NULL, identity permutation is taken.}

\item{delta}{hyper-parameter of a Bayesian model. Has to be bigger than 2.}

\item{D_matrix}{hyper-parameter of a Bayesian model. Square matrix of the same size as \code{U}. When NULL, the identity matrix is taken.}

\item{show_progress_bar}{boolean, indicate weather or not show the progress bar.}
}
\value{
list of 3 items: \code{acceptance_rate}, \code{goal_function_values}, \code{points}
}
\description{
Uses Metropolis-Hastings algorithm to find the permutation that maximizes the likelihood of observed data.
}
\examples{
perm_size <- 6
mu <- numeric(perm_size)
# sigma is a permutation (1,2,3,4,5,6)
sigma <- matrix(data = c(1.0, 0.8, 0.6, 0.4, 0.6, 0.8,
                         0.8, 1.0, 0.8, 0.6, 0.4, 0.6,
                         0.6, 0.8, 1.0, 0.8, 0.6, 0.4,
                         0.4, 0.6, 0.8, 1.0, 0.8, 0.6,
                         0.6, 0.4, 0.6, 0.8, 1.0, 0.8,
                         0.8, 0.6, 0.4, 0.6, 0.8, 1.0),
                nrow=perm_size, byrow = TRUE)
n_number <- 6
Z <- MASS::mvrnorm(n_number, mu = mu, Sigma = sigma)
U <- (t(Z) \%*\% Z)/n_number
start <- permutations::id
mh <- MH(U=U, n_number=10, max_iter=100, start=start,
         delta=3, D_matrix=diag(nrow=dim(U)[1]))
}
