% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_functions.R
\name{plot.gips}
\alias{plot.gips}
\title{Plot optimization convergence}
\usage{
\method{plot}{gips}(
  x,
  type = "both",
  logarithmic = TRUE,
  color = NULL,
  title_text = "Convergence plot",
  xlabel = "number of function calls",
  ylabel = NULL,
  show_legend = TRUE,
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{Object of class gips.}

\item{type}{Character. A type of a plot. Either "all", "best" or "both". For "all" plots values of goal function for all visited state. For "best" the best value of goal function up to the point are plotted. For "both" both lines are plotted.}

\item{logarithmic}{boolean.}

\item{color}{Vector of olors to be used to plot lines.}

\item{title_text}{Text to be in a title of the plot.}

\item{xlabel}{Text to be on the bottom of the plot.}

\item{ylabel}{Text to be on the left of the plot.}

\item{show_legend}{boolean.}

\item{ylim}{Limits of y axis. When \code{NULL}, the minimum and maximum of the \code{\link{log_likelihood_of_perm}} is taken.}

\item{...}{additional arguments passed to \code{\link{print}}.}
}
\value{
Invisible NULL.
}
\description{
Plot method for gips objects.
}
