% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_class.R
\name{plot.gips}
\alias{plot.gips}
\title{Plot optimized matrix or optimization convergence}
\usage{
\method{plot}{gips}(
  x,
  type = NA,
  logarithmic_y = TRUE,
  logarithmic_x = FALSE,
  color = NULL,
  title_text = "Convergence plot",
  xlabel = NULL,
  ylabel = NULL,
  show_legend = TRUE,
  ylim = NULL,
  xlim = NULL,
  ...
)
}
\arguments{
\item{x}{Object of class gips. Has to first be optimized with \code{\link[gips]{find_MAP}}.}

\item{type}{Character. A type of a plot. one of \code{c("heatmap", "all", "best", "both")}. For "heatmap", plots a heatmap of the \code{S} matrix inside the \code{gips} object that was projected on the permutation in the \code{gips} object. For "all", plots the line of the posterioris for all visited state. For "best", plots the line of the biggest posterioris up to the moment For "both", both lines from "all" and "best" are plotted. Default value is \code{NA}, which will be changed to "heatmap" for non-optimized \code{gips} objects, and to "both" for optimized ones. Using the default produces a warning. For the \code{type = "heatmap"}, all other arguments are ignored.}

\item{logarithmic_y}{boolean.}

\item{logarithmic_x}{boolean.}

\item{color}{Vector of colors to be used to plot lines.}

\item{title_text}{Text to be in a title of the plot.}

\item{xlabel}{Text to be on the bottom of the plot.}

\item{ylabel}{Text to be on the left of the plot.}

\item{show_legend}{boolean.}

\item{ylim}{Limits of y axis. When \code{NULL}, the minimum and maximum of the \code{\link[gips]{log_posteriori_of_gips}} is taken.}

\item{xlim}{Limits of x axis. When \code{NULL}, the whole optimization process is shown.}

\item{...}{Additional arguments passed to \code{\link[stats]{heatmap}} or other various elements of the plot.}
}
\value{
Invisible NULL.
}
\description{
Plot method for \code{gips} object.
}
