% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_S3_functions.R
\name{plot.gips}
\alias{plot.gips}
\title{Plot optimization convergence}
\usage{
\method{plot}{gips}(
  x,
  logarithmic = TRUE,
  title_text = "Convergence plot",
  xlabel = "number of function calls",
  ylabel = NULL,
  show_legend = TRUE,
  ...
)
}
\arguments{
\item{x}{Object of class gips. Output of MH function.}

\item{logarithmic}{boolean.}

\item{title_text}{Text to be in a title of the plot.}

\item{xlabel}{Text to be on the bottom of the plot.}

\item{ylabel}{Text to be on the left of the plot.}

\item{show_legend}{boolean.}

\item{...}{additional arguments passed to \code{\link{print}}.}
}
\value{
Invisible NULL.
}
\description{
Plot method for gips objects returned by \code{\link{MH}}.
}
