% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_orthogonal_matrix.R
\name{prepare_orthogonal_matrix}
\alias{prepare_orthogonal_matrix}
\title{Prepare orthogonal matrix}
\usage{
prepare_orthogonal_matrix(perm, perm_size, basis = NULL)
}
\arguments{
\item{perm}{permuations::cycle object}

\item{perm_size}{size of permutation}

\item{basis}{matrix with basis vectors in COLUMNS. Identity by default}
}
\value{
matrix p x p with columns from V object elements, sorted according to
Theorem 6.
}
\description{
Calculate orthogonal matrix U_Gamma
for decomposition in Theorem 1. To calculate it, we use Theorem 6.
}
