% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_orthogonal_matrix.R
\name{prepare_orthogonal_matrix}
\alias{prepare_orthogonal_matrix}
\title{Prepare orthogonal matrix}
\usage{
prepare_orthogonal_matrix(perm, perm_size = NULL, basis = NULL)
}
\arguments{
\item{perm}{Object of class \code{gips_perm} or \code{permutations::cycle}.}

\item{perm_size}{Size of permutation. Required if \code{perm} is of \code{permutations::cycle} class.}

\item{basis}{A matrix with basis vectors in COLUMNS. Identity by default.}
}
\value{
A matrix p x p with columns from V object elements, sorted according to
Theorem 6.
}
\description{
Calculate orthogonal matrix U_Gamma
for decomposition in Theorem 1. To calculate it, we use Theorem 6.
}
\examples{
gperm <- gips_perm("(1,2,3)(4,5)", 6)
U_g <- prepare_orthogonal_matrix(gperm)
}
