% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_S3_functions.R
\name{print.gips}
\alias{print.gips}
\title{Printing gips object}
\usage{
\method{print}{gips}(x, ...)
}
\arguments{
\item{x}{object of class gips. Output of MH function.}

\item{...}{additional arguments passed to \code{\link{cat}}.}
}
\value{
Invisible NULL.
}
\description{
Printing function for gips class.
}
