% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_class.R
\name{print.gips}
\alias{print.gips}
\title{Printing \code{gips} object}
\usage{
\method{print}{gips}(x, log_value = TRUE, digits = Inf, ...)
}
\arguments{
\item{x}{An object of \code{gips} class.}

\item{log_value}{A logical. Whether to print
the exp of a value of a \code{\link[=log_posteriori_of_gips]{log_posteriori_of_gips()}}
or leave it in logarithmic form.}

\item{digits}{A number of digits after the comma
for a posteriori to be presented. It can be negative.
By default, \code{Inf}. It is passed to \code{\link[base:Round]{base::round()}}.}

\item{...}{An additional arguments passed to \code{\link[base:cat]{base::cat()}}.}
}
\value{
Invisible NULL.
}
\description{
Printing function for \code{gips} class.
}
