% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_class.R
\name{print.gips}
\alias{print.gips}
\title{Printing \code{gips} object}
\usage{
\method{print}{gips}(x, log_value = TRUE, digits = Inf, ...)
}
\arguments{
\item{x}{An object of class \code{gips}.}

\item{log_value}{A logical. Weather to print the exp of a value of a \code{\link[gips]{log_posteriori_of_gips}} or leave it in logarithmic form.}

\item{digits}{A number of digits after the comma for posteriori to be presented. Can be negative. Be default, \code{Inf}. It is passed to \code{\link[base]{round}}.}

\item{...}{An additional arguments passed to \code{\link[base]{cat}}.}
}
\value{
Invisible NULL.
}
\description{
Printing function for \code{gips} class.
}
