% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_class.R
\name{print.gips}
\alias{print.gips}
\title{Printing \code{gips} object}
\usage{
\method{print}{gips}(x, log_value = TRUE, digits = Inf, ...)
}
\arguments{
\item{x}{object of class \code{gips}.}

\item{log_value}{logical. Weather to print the exp of a value of a \code{\link{log_likelihood_of_gips}} or leave it in logarithmic form.}

\item{digits}{Number of digits after the comma for likelihood to be presented. Can be negative. Be default, \code{Inf}. It is passed to \code{\link{round}}.}

\item{...}{additional arguments passed to \code{\link{cat}}.}
}
\value{
Invisible NULL.
}
\description{
Printing function for \code{gips} class.
}
