% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gips_perm_class.R
\name{print.gips_perm}
\alias{print.gips_perm}
\title{Printing \code{gips_perm} object}
\usage{
\method{print}{gips_perm}(x, ...)
}
\arguments{
\item{x}{An object of a \code{gips_perm} class.}

\item{...}{Further arguments passed to \code{\link[permutations:print]{permutations::print.cycle()}}.}
}
\value{
Returns its argument invisibly, after printing it.
}
\description{
Printing function for a \code{gips_perm} class.
}
\examples{
g_perm <- gips_perm(permutations::as.cycle("(5,4)"), 5)
print(g_perm)
}
