% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_matrix.R
\name{project_matrix}
\alias{project_matrix}
\title{Project matrix}
\usage{
project_matrix(U, perm, perm_size)
}
\arguments{
\item{U}{matrix to be projected}

\item{perm}{permutation. Generator of a permutation group}

\item{perm_size}{size of permutation}
}
\value{
projected matrix
}
\description{
Project matrix on the space of symmetrical matrices invariant
by a cyclic group of permutations.
}
