% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_matrix.R
\name{project_matrix}
\alias{project_matrix}
\title{Project matrix}
\usage{
project_matrix(S, perm, precomputed_equal_indices = NULL)
}
\arguments{
\item{S}{matrix to be projected.}

\item{perm}{permutation. Generator of a permutation group.
Either of \code{gips_perm} or \code{permutations::cycle} class.}

\item{precomputed_equal_indices}{used in internal calculations in case when the equal indices have already been calculated; If it is not the case, leave this parameter as \code{NULL} and those will be computed}
}
\value{
projected matrix
}
\description{
Project matrix on the space of symmetrical matrices invariant
by a cyclic group of permutations.
}
\examples{
gperm <- gips_perm(permutations::as.word(c(4,3,2,1,5)), 7)  # permutation (1,4)(2,3)(5)(6)(7)
S <- matrix(rnorm(49), nrow = 7)
projected_S <- project_matrix(S, perm = gperm)
}
