% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_matrix.R
\name{project_matrix}
\alias{project_matrix}
\title{Project matrix}
\usage{
project_matrix(U, perm, perm_size, precomputed_equal_indices = NULL)
}
\arguments{
\item{U}{matrix to be projected}

\item{perm}{permutation. Generator of a permutation group}

\item{perm_size}{size of permutation}

\item{precomputed_equal_indices}{used in internal calculations in case when the equal indices have already been calculated; If it is not the case, leave this parameter as \code{NULL} and those will be computed}
}
\value{
projected matrix
}
\description{
Project matrix on the space of symmetrical matrices invariant
by a cyclic group of permutations.
}
\examples{
project_matrix(U = matrix(rnorm(49), nrow = 7),
                         perm = permutations::as.cycle(permutations::as.word(c(4,3,2,1,5))),
                         perm_size = 7)
}
